pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "se-cvla"
version = "0.1.0"
description = "Self-Evolving Causal Vision-Language-Action Model for Autonomous Driving"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "SE-CVLA Research Team" }]

dependencies = [
    # Core deep learning
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "einops>=0.7.0",
    "timm>=1.0.0",

    # Transformers / LLM backbone
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "peft>=0.10.0",                   # LoRA for efficient fine-tuning

    # Causal / graph libraries
    "torch-geometric>=2.5.0",         # GNN for SCM
    "networkx>=3.3",                   # Causal graph manipulation
    "pgmpy>=0.1.25",                   # Probabilistic graphical models
    "dowhy>=0.11",                     # DoWhy causal inference

    # Diffusion (trajectory generation — matches Alpamayo 1.5 architecture)
    "diffusers>=0.27.0",

    # Uncertainty quantification
    "uncertainty-toolbox>=0.1.1",
    "torch-uncertainty>=0.2.0",

    # Data / HuggingFace
    "datasets>=2.19.0",
    "huggingface-hub>=0.23.0",
    "physical-ai-av>=0.1.0",          # NVIDIA PhysicalAI-AV package

    # Training infrastructure
    "pytorch-lightning>=2.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.17.0",
    "tensorboard>=2.17.0",

    # Evaluation
        "scikit-learn>=1.5.0",
    "scipy>=1.13.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # Visualization
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "plotly>=5.22.0",
    "pyvis>=0.3.2",                   # Interactive causal graph visualization

    # Simulation interfaces
    "nuplan-devkit>=1.3.0",           # nuPlan closed-loop simulator
    "gymnasium>=0.29.0",              # OpenAI Gymnasium interface

    # Attention
    "flash-attn>=2.5.0",

    # Utilities
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
carla = ["carla>=0.9.15"]            # CARLA simulator (optional)
dev   = ["ruff", "mypy", "pre-commit", "ipykernel", "jupyter"]

[project.scripts]
se-cvla-train    = "scripts.train:main"
se-cvla-eval     = "scripts.evaluate:main"
se-cvla-sim      = "scripts.closed_loop_eval:main"

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "--cov=se_cvla --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = ["models", "data", "training", "evaluation", "simulation", "scripts"]
